watchlist-service/requirements.txt
Django>=4.2
djangorestframework>=3.14
PyJWT>=2.8
gunicorn>=20.1
dj-database-url>=1.0
psycopg2-binary>=2.9
python-dotenv>=1.0